% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bfs_data.R
\name{read_BFS_data}
\alias{read_BFS_data}
\title{Get BFS Data from SwitchDrive}
\usage{
read_BFS_data(
  survey = "EHA",
  first_wave = T,
  year = 2023,
  cohort = F,
  HSG_only = T
)
}
\arguments{
\item{survey}{Specify the survey type you wish to load. Can be either "EHA" or "SSEE".}

\item{first_wave}{For "EHA" survey please specify the survey wave.
If TRUE, "Erstbefragung" is used. If FALSE, it will search for "Zweitbefragung".}

\item{year}{Specify survey year. By default, it is set to 2023.
If cohort year is used, set cohort = TRUE. Cohort option is only available for EHA.}

\item{cohort}{By default FALSE. Set to TRUE if year is the cohort year.}

\item{HSG_only}{By default TRUE. If TRUE, only respondents affiliated to HSG are returned.
If all Swiss universities are to be included, set to FALSE.}
}
\value{
Function returns dataframe of the specified BFS survey.
}
\description{
This function retrieves the BFS Data that are stored on SwitchDrive and returns them as a dataframe.
}
\examples{
 eha_23_eb_all <- read_BFS_data(HSG_only = FALSE)
 ssee_24_hsg <- read_BFS_data(survey = "SSEE", year = 2024)

}
