% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sd.R
\name{read_sd}
\alias{read_sd}
\title{Import any data files on SwitchDrive}
\usage{
read_sd(path, delim = "comma", ...)
}
\arguments{
\item{path}{Provide SwitchDrive path to the designated file that needs to be imported.}

\item{delim}{For csv-files, specify the deliminator. By default "comma", alternatively "semicol".}

\item{...}{Pass parameters for tidyverse import functionss}
}
\value{
Dataframe
}
\description{
Function allows to read any data files on SwitchDrive, by using tidyverse import functions.
}
\examples{
# SAV
df <- sd_read("BFS/EHA/EHA_2023_HSG.sav")

df <- sd_read("Projekte/umfrage_2024.xlsx", sheet = "Rohdaten")

df <- sd_read("Daten/export.csv", delim = "semicol")

df <- sd_read("Daten/export.csv")


}
