% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_connection.R
\name{sd_connect}
\alias{sd_connect}
\title{Connect to SwitchDrive}
\usage{
sd_connect(
  username = Sys.getenv("SWITCH_USER"),
  password = Sys.getenv("SWITCH_PASSWORD")
)
}
\arguments{
\item{username}{Username is taken by default from .Renviron.
If no username is available, it can be passed directly in function or use sd_setup() instead.}

\item{password}{Password is taken by default from .Renviron.
If no Password is available, it can be passed directly in function or use sd_setup() instead.}
}
\value{
List of credentials: Username, Password and Path.
}
\description{
This function connects to a SwitchDrive Account and returns all relevant credentials.
Credentials are stored, so that connection only has to be established once.
}
\examples{
sd_connect()
}
