% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_connection.R
\name{sd_get_connection}
\alias{sd_get_connection}
\title{Helper function to establish connection}
\usage{
sd_get_connection()
}
\description{
Used as helper function in functions that retrieve SwitchDrive data.
}
