% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_list_files.R
\name{sd_list_files}
\alias{sd_list_files}
\title{Get list of files and folder on SwitchDrive}
\usage{
sd_list_files(path = "")
}
\arguments{
\item{path}{Set path of folder the contents should be extracted from.
If none is provided, the function will return the contents of root.
Only set path from root, full url will be provided by the function.}
}
\value{
Vector of files and folders contained within the set path.
Folders are indicated with a "/" at the end of the string.
}
\description{
Use this function to get a list all the files and folders within a certain path.
}
\examples{
#Get root list
sd_list_files()

#list of all working files of SSEE
sd_list_files("BFS Daten/02 SSEE Studenten/01 Arbeitsdaten")

}
