% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd_setup.R
\name{sd_setup}
\alias{sd_setup}
\title{Set SwitchDrive Credentials}
\usage{
sd_setup()
}
\description{
This function allows you to set your credentials of SwitchDrive.
Call this function when you have never connected to your SwitchDrive Account via R, to store username and password for all connections.
It is advised to set a app-specific password.
}
\examples{
sd_setup()


}
